Develop (#52)
* Don't consume all arguments after --extensions

This behavior is now more like other utilities where specifying the flag
multiple times extends the value of the argument.  For example,
    -e '.nc .h5 .zip'
becomes
    -e '.nc' -e '.h5' -e '.zip'

This is less fragile for the user and possibly less confusing how the
argument should be formatted on the command line.

* Add ability to execute arbitrary commands on each downloaded file

I did it this way so each file could be compressed without hard-coding
the compression algorithm. But I could see this being used to run a
pre-processing script on each downloaded file.

* updated README and tests for additive -e examples

* force 'action'

* merged code for extensions, process call, and updated documentation

* fix for https://github.com/podaac/data-subscriber/issues/28

* updated CHANGELOG

* Issue 33 (#35)

* Develop (#32)

* Don't consume all arguments after --extensions

This behavior is now more like other utilities where specifying the flag
multiple times extends the value of the argument.  For example,
    -e '.nc .h5 .zip'
becomes
    -e '.nc' -e '.h5' -e '.zip'

This is less fragile for the user and possibly less confusing how the
argument should be formatted on the command line.

* Add ability to execute arbitrary commands on each downloaded file

I did it this way so each file could be compressed without hard-coding
the compression algorithm. But I could see this being used to run a
pre-processing script on each downloaded file.

* updated README and tests for additive -e examples

* force 'action'

* merged code for extensions, process call, and updated documentation

* fix for https://github.com/podaac/data-subscriber/issues/28

* updated CHANGELOG

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>

* closes #33
added 'files to download' to non-verbose output

* updated changelog

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>

* updated changelog to reflect accurate versions released

* version bump for master delivery

* Bulk download capability (#45)

* added pypi release badge to README

* began separating out functionality for data downloader

* 1.7.2 merge into develop (#43)

* Don't consume all arguments after --extensions

This behavior is now more like other utilities where specifying the flag
multiple times extends the value of the argument.  For example,
    -e '.nc .h5 .zip'
becomes
    -e '.nc' -e '.h5' -e '.zip'

This is less fragile for the user and possibly less confusing how the
argument should be formatted on the command line.

* Add ability to execute arbitrary commands on each downloaded file

I did it this way so each file could be compressed without hard-coding
the compression algorithm. But I could see this being used to run a
pre-processing script on each downloaded file.

* updated README and tests for additive -e examples

* force 'action'

* merged code for extensions, process call, and updated documentation

* fix for https://github.com/podaac/data-subscriber/issues/28

* updated CHANGELOG

* Issue 33 (#35)

* Develop (#32)

* Don't consume all arguments after --extensions

This behavior is now more like other utilities where specifying the flag
multiple times extends the value of the argument.  For example,
    -e '.nc .h5 .zip'
becomes
    -e '.nc' -e '.h5' -e '.zip'

This is less fragile for the user and possibly less confusing how the
argument should be formatted on the command line.

* Add ability to execute arbitrary commands on each downloaded file

I did it this way so each file could be compressed without hard-coding
the compression algorithm. But I could see this being used to run a
pre-processing script on each downloaded file.

* updated README and tests for additive -e examples

* force 'action'

* merged code for extensions, process call, and updated documentation

* fix for https://github.com/podaac/data-subscriber/issues/28

* updated CHANGELOG

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>

* closes #33
added 'files to download' to non-verbose output

* updated changelog

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>

* updated changelog to reflect accurate versions released

* version bump for master delivery

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>

* fixed tests

* removed erroneous printouts for pytest fixes

* fixed some bugs
closes #39
closes #40

* Update README.md

* updates for flake8

* fixed issue with time_offset variables, added manual tests

* flake8 updates for access, subscriber

* added downloader tests, flake8 updates for download client

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>

* updated tests

* fixed path issue for built python command line tooling

* Ignore error if directory already exists (#47)

* Ignore error if directory already exists

Fixes #46

* issues/46: Ignore errors if destination directory already exists.

Co-authored-by: Frank Greguska <Francis.Greguska@jpl.nasa.gov>

* Issues/44 (#48)

* Added functionality to name .update file to .update__<COLLECTION> so that the same directory can be re-used.
Added tests to fix this functionality
closes #44

* flake8 fixes

* added updates to .update file

* Issues/41 (#50)

* added cycle based downloads

* udpated changelog

* 180 docs (#51)

* added changelog info for 1.80, created initial markdown files

* added subsriber/downloader links

* added documentation

* Added downloader/subscriber specific docs.
fixed some issues with imports

* Update README.md

Co-authored-by: Joe Sapp <joe.sapp@noaa.gov>
Co-authored-by: mgangl <mike.gangl@gmail.com>
Co-authored-by: Frank Greguska <89428916+frankinspace@users.noreply.github.com>
Co-authored-by: Frank Greguska <Francis.Greguska@jpl.nasa.gov>